Split Tooltip into provider/controller
