+++ lib/Alchemy/Phrasea/Application.php
@@ -12,7 +12,6 @@
 namespace Alchemy\Phrasea;
 
 use Alchemy\Geonames\GeonamesServiceProvider;
-use Alchemy\Phrasea\ControllerProvider\Prod\Tooltip;
 use Alchemy\Phrasea\ControllerProvider\Prod\Upload;
 use Alchemy\Phrasea\ControllerProvider\Prod\UsrLists;
 use Alchemy\Phrasea\ControllerProvider\Prod\WorkZone;
@@ -314,6 +313,7 @@ class Application extends SilexApplication
             'Alchemy\Phrasea\ControllerProvider\Prod\Share' => [],
             'Alchemy\Phrasea\ControllerProvider\Prod\Story' => [],
             'Alchemy\Phrasea\ControllerProvider\Prod\Tools' => [],
+            'Alchemy\Phrasea\ControllerProvider\Prod\Tooltip' => [],
             'Alchemy\Phrasea\ControllerProvider\Prod\TOU' => [],
             'Alchemy\Phrasea\ControllerProvider\Datafiles' => [],
             'Alchemy\Phrasea\ControllerProvider\Lightbox' => [],
@@ -627,7 +627,6 @@ class Application extends SilexApplication
 
         $this->mount('/prod/WorkZone', new WorkZone());
         $this->mount('/prod/lists', new UsrLists());
-        $this->mount('/prod/tooltip', new Tooltip());
         $this->mount('/prod/upload/', new Upload());
 
         $this->mount('/user/preferences/', new Preferences());
@@ -680,6 +679,7 @@ class Application extends SilexApplication
             '/prod/share/'                 => 'Alchemy\Phrasea\ControllerProvider\Prod\Share',
             '/prod/story'                  => 'Alchemy\Phrasea\ControllerProvider\Prod\Story',
             '/prod/tools/'                 => 'Alchemy\Phrasea\ControllerProvider\Prod\Tools',
+            '/prod/tooltip'                => 'Alchemy\Phrasea\ControllerProvider\Prod\Tooltip',
             '/prod/TOU/'                   => 'Alchemy\Phrasea\ControllerProvider\Prod\TOU',
             '/prod/'                       => 'Alchemy\Phrasea\ControllerProvider\Prod\Root',
             '/setup'                       => 'Alchemy\Phrasea\ControllerProvider\Setup',
